no functional change. move interventions API tests from test_views to test_api_interventions.
